Initial proof of concept.
The web server runs in a Docker container.
It handles OAuth request from OnShape and presents content within
the OnShape UI as a right side element panel in a Part Studio.
When the panel is opened, it retrieves the index of CAM files from
the active Part Studio and displays the contents of the first one.