Refactor .env to support running test instances.
Instead of having the .env be loaded by the application at runtime,
its properties are injected into the container by Docker. That change
also makes it easier to run test instances (e.g. on different ports)
by parameterizing the contents of the docker-compose.yaml configuration.

Added the icons to the repository with attribution.

Updated the documentation.