Detect when the browser is blocking third-party cookies.
OnShape embeds Camel as a panel by loading it into an iframe. Because
the top-level document's origin is different from the Camel website,
the browser considers Camel's session cookie to be a third-party cookie.
Depending on the browser's configuration and behavior, it may block
Camel from accessing its cookies when embedded in a frame.

As a workaround, offers the user the choice of trying to login again
or to open Camel on its own in a new tab where it will be considered
to be the first-party as far as cookies are concerned.